added S1 group and tests
